Common, Engine: picked out RoomData class from RoomStruct
RoomData is practically what RoomStruct was previously, a collection of all the room data read from the compiled room file.
The main difference is that all the Bitmap fields are replaced with PixelBuffer, which is a much simpler class, contains only pixel data without any additional dependencies (such as drawing operations etc).
This lets to include RoomData and related reading/writing function without adding full dependency on Allegro4 library.

RoomStruct inherits RoomData, and adds Bitmap fields that are easier to use at runtime, or whenever we must use room backgrounds and masks as images in drawing operations.
