+++ Common/CMakeLists.txt
@@ -88,6 +88,8 @@ target_sources(common
     game/room_file_base.cpp
     game/room_file_deprecated.cpp
     game/room_version.h
+    game/roomdata.cpp
+    game/roomdata.h
     game/roomstruct.cpp
     game/roomstruct.h
     gfx/allegrobitmap.cpp
