+++ Common/game/roomstruct.cpp
@@ -11,10 +11,8 @@
 // https://opensource.org/license/artistic-2-0/
 //
 //=============================================================================
-
-#include "ac/common.h" // quit
-#include "game/room_file.h"
 #include "game/roomstruct.h"
+#include "game/room_version.h"
 #include "gfx/bitmap.h"
 
 namespace AGS
@@ -22,176 +20,31 @@ namespace AGS
 namespace Common
 {
 
-RoomOptions::RoomOptions()
-    : StartupMusic(0)
-    , SaveLoadDisabled(false)
-    , PlayerCharOff(false)
-    , PlayerView(0)
-    , MusicVolume(kRoomVolumeNormal)
-    , Flags(0)
-{
-}
-
-RoomBgFrame::RoomBgFrame()
-    : IsPaletteShared(false)
-{
-    memset(Palette, 0, sizeof(Palette));
-}
-
-RoomEdges::RoomEdges()
-    : Left(0)
-    , Right(0)
-    , Top(0)
-    , Bottom(0)
-{
-}
-
-RoomEdges::RoomEdges(int l, int r, int t, int b)
-    : Left(l)
-    , Right(r)
-    , Top(t)
-    , Bottom(b)
-{
-}
-
-RoomObjectInfo::RoomObjectInfo()
-    : Room(-1)
-    , X(0)
-    , Y(0)
-    , Sprite(0)
-    , IsOn(false)
-    , Baseline(0)
-    , Transparency(0)
-    , Flags(0)
-{
-}
-
-RoomRegion::RoomRegion()
-    : Light(0)
-    , Tint(0)
-{
-}
-
-WalkArea::WalkArea()
-    : CharacterView(0)
-    , ScalingFar(0)
-    , ScalingNear(NOT_VECTOR_SCALED)
-    , PlayerView(0)
-    , Top(-1)
-    , Bottom(-1)
-{
-}
-
-WalkBehind::WalkBehind()
-    : Baseline(0)
-{
-}
-
-MessageInfo::MessageInfo()
-    : DisplayAs(0)
-    , Flags(0)
-{
-}
-
-RoomStruct::RoomStruct()
-{
-    InitDefaults();
-}
-
 RoomStruct::~RoomStruct()
 {
     Free();
 }
 
-void RoomStruct::Free()
+void RoomStruct::InitRuntimeData()
 {
     for (size_t i = 0; i < (size_t)MAX_ROOM_BGFRAMES; ++i)
-        BgFrames[i].Graphic.reset();
+        BgImages[i].reset(new Bitmap(std::move(BgFrames[i].GraphicBuf)));
+    HotspotMask.reset(new Bitmap(std::move(HotspotMaskBuf)));
+    RegionMask.reset(new Bitmap(std::move(RegionMaskBuf)));
+    WalkAreaMask.reset(new Bitmap(std::move(WalkAreaMaskBuf)));
+    WalkBehindMask.reset(new Bitmap(std::move(WalkBehindMaskBuf)));
+}
+
+void RoomStruct::Free()
+{
+    RoomData::Free();
+
+    for (size_t i = 0; i < (size_t)MAX_ROOM_BGFRAMES; ++i)
+        BgImages[i].reset();
     HotspotMask.reset();
     RegionMask.reset();
     WalkAreaMask.reset();
     WalkBehindMask.reset();
-
-    LocalVariables.clear();
-    Interaction.reset();
-    Properties.clear();
-    for (size_t i = 0; i < (size_t)MAX_ROOM_HOTSPOTS; ++i)
-    {
-        Hotspots[i].Interaction.reset();
-        Hotspots[i].Properties.clear();
-    }
-    Objects.clear();
-    for (size_t i = 0; i < (size_t)MAX_ROOM_REGIONS; ++i)
-    {
-        Regions[i].Interaction.reset();
-        Regions[i].Properties.clear();
-    }
-
-    FreeMessages();
-    FreeScripts();
-}
-
-void RoomStruct::FreeMessages()
-{
-    for (uint32_t i = 0; i < MessageCount; ++i)
-    {
-        Messages[i].Free();
-        MessageInfos[i] = MessageInfo();
-    }
-    MessageCount = 0;
-}
-
-void RoomStruct::FreeScripts()
-{
-    CompiledScript.reset();
-
-    EventHandlers.reset();
-    for (uint32_t i = 0; i < HotspotCount; ++i)
-        Hotspots[i].EventHandlers.reset();
-    for (auto &obj : Objects)
-        obj.EventHandlers.reset();
-    for (uint32_t i = 0; i < RegionCount; ++i)
-        Regions[i].EventHandlers.reset();
-}
-
-void RoomStruct::InitDefaults()
-{
-    DataVersion     = kRoomVersion_Current;
-    GameID          = NO_GAME_ID_IN_ROOM_FILE;
-
-    _legacyResolution = kRoomResolution_Real;
-    MaskResolution  = 1;
-    Width           = 320;
-    Height          = 200;
-
-    Options         = RoomOptions();
-    Edges           = RoomEdges(0, 317, 40, 199);
-
-    BgFrameCount    = 1;
-    HotspotCount    = 0;
-    RegionCount     = 0;
-    WalkAreaCount   = 0;
-    WalkBehindCount = 0;
-    MessageCount    = 0;
-
-    for (size_t i = 0; i < (size_t)MAX_ROOM_HOTSPOTS; ++i)
-        Hotspots[i] = RoomHotspot();
-    for (size_t i = 0; i < (size_t)MAX_ROOM_REGIONS; ++i)
-        Regions[i] = RoomRegion();
-    for (size_t i = 0; i < (size_t)MAX_WALK_AREAS; ++i)
-        WalkAreas[i] = WalkArea();
-    for (size_t i = 0; i < (size_t)MAX_WALK_BEHINDS; ++i)
-        WalkBehinds[i] = WalkBehind();
-    
-    BackgroundBPP = 1;
-    BgAnimSpeed = 5;
-
-    memset(Palette, 0, sizeof(Palette));
-}
-
-void RoomStruct::SetLegacyResolution(RoomResolutionType resolution)
-{
-    _legacyResolution = resolution;
 }
 
 Bitmap *RoomStruct::GetMask(RoomAreaMask mask) const
@@ -205,6 +58,7 @@ Bitmap *RoomStruct::GetMask(RoomAreaMask mask) const
     default: return nullptr;
     }
 }
+
 void RoomStruct::SetMask(RoomAreaMask mask, std::unique_ptr<Bitmap> &&bmp)
 {
     switch (mask)
@@ -227,47 +81,6 @@ void RoomStruct::CopyMask(RoomAreaMask mask, const Bitmap *bitmap)
     }
 }
 
-float RoomStruct::GetMaskScale(RoomAreaMask mask) const
-{
-    switch (mask)
-    {
-    case kRoomAreaWalkBehind: return 1.f; // walk-behinds always 1:1 with room size
-    case kRoomAreaHotspot:
-    case kRoomAreaWalkable:
-    case kRoomAreaRegion:
-        return 1.f / MaskResolution;
-    default: return 0.f;
-    }
-}
-
-bool RoomStruct::HasRegionLightLevel(int id) const
-{
-    if (id >= 0 && id < MAX_ROOM_REGIONS)
-        return Regions[id].Tint == 0;
-    return false;
-}
-
-bool RoomStruct::HasRegionTint(int id) const
-{
-    if (id >= 0 && id < MAX_ROOM_REGIONS)
-        return Regions[id].Tint != 0;
-    return false;
-}
-
-int RoomStruct::GetRegionLightLevel(int id) const
-{
-    if (id >= 0 && id < MAX_ROOM_REGIONS)
-        return HasRegionLightLevel(id) ? Regions[id].Light : 0;
-    return 0;
-}
-
-int RoomStruct::GetRegionTintLuminance(int id) const
-{
-    if (id >= 0 && id < MAX_ROOM_REGIONS)
-        return HasRegionTint(id) ? (Regions[id].Light * 10) / 25 : 0;
-    return 0;
-}
-
 
 PBitmap FixBitmap(PBitmap bmp, int width, int height)
 {
@@ -282,7 +95,7 @@ void UpscaleRoomBackground(RoomStruct *room, bool game_is_hires)
     if (room->DataVersion >= kRoomVersion_303b || !game_is_hires)
         return;
     for (size_t i = 0; i < room->BgFrameCount; ++i)
-        room->BgFrames[i].Graphic = FixBitmap(room->BgFrames[i].Graphic, room->Width, room->Height);
+        room->BgImages[i] = FixBitmap(room->BgImages[i], room->Width, room->Height);
     FixRoomMasks(room);
 }
 
@@ -290,7 +103,7 @@ void FixRoomMasks(RoomStruct *room)
 {
     if (room->MaskResolution <= 0)
         return;
-    Bitmap *bkg = room->BgFrames[0].Graphic.get();
+    Bitmap *bkg = room->BgImages[0].get();
     if (bkg == nullptr)
         return;
     // TODO: this issue is somewhat complicated. Original code was relying on
