+++ Solutions/Common.Lib/Common.Lib.vcxproj
@@ -634,6 +634,7 @@
     <ClCompile Include="..\..\Common\game\interactions.cpp" />
     <ClCompile Include="..\..\Common\game\interactions_deprecated.cpp" />
     <ClCompile Include="..\..\Common\game\main_game_file.cpp" />
+    <ClCompile Include="..\..\Common\game\roomdata.cpp" />
     <ClCompile Include="..\..\Common\game\roomstruct.cpp" />
     <ClCompile Include="..\..\Common\game\room_file.cpp" />
     <ClCompile Include="..\..\Common\game\room_file_base.cpp" />
@@ -824,6 +825,7 @@
     <ClInclude Include="..\..\Common\game\interactions_deprecated.h" />
     <ClInclude Include="..\..\Common\game\main_game_file.h" />
     <ClInclude Include="..\..\Common\game\plugininfo.h" />
+    <ClInclude Include="..\..\Common\game\roomdata.h" />
     <ClInclude Include="..\..\Common\game\roomstruct.h" />
     <ClInclude Include="..\..\Common\game\room_version.h" />
     <ClInclude Include="..\..\Common\game\room_file.h" />
